Move the setUpCryptoservices to its own file.
